Refactor and rename Hamiltonian to SpatialHamiltonian
